clients/requirements.txt
paho-mqtt==1.3.0
pkg-resources==0.0.0
